Updated existing resource loaders to resourceLoaderJson. Renamed package
