+++ wolfyutilities/src/main/java/me/wolfyscript/utilities/main/WUPlugin.java
@@ -65,8 +65,8 @@ import me.wolfyscript.utilities.main.commands.InputCommand;
 import me.wolfyscript.utilities.main.commands.SpawnParticleAnimationCommand;
 import me.wolfyscript.utilities.main.commands.SpawnParticleEffectCommand;
 import me.wolfyscript.utilities.main.configs.WUConfig;
-import me.wolfyscript.utilities.main.expansions.ResourceLoaderParticleAnimations;
-import me.wolfyscript.utilities.main.expansions.ResourceLoaderParticleEffects;
+import me.wolfyscript.utilities.main.resource_loader.ResourceLoaderParticleAnimations;
+import me.wolfyscript.utilities.main.resource_loader.ResourceLoaderParticleEffects;
 import me.wolfyscript.utilities.main.listeners.BlockListener;
 import me.wolfyscript.utilities.main.listeners.EquipListener;
 import me.wolfyscript.utilities.main.listeners.GUIInventoryListener;
